+++ pkg/dbt_cloud/snowflake_credential.go
@@ -14,7 +14,7 @@ type SnowflakeCredentialResponse struct {
 
 type SnowflakeCredential struct {
 	ID                   *int   `json:"id"`
-	Account_Id           int    `json:"account_id"`
+	Account_Id           int64  `json:"account_id"`
 	Project_Id           int    `json:"project_id"`
 	Type                 string `json:"type"`
 	State                int    `json:"state"`
