+++ pkg/dbt_cloud/starburst_credential.go
@@ -25,7 +25,7 @@ type StarburstUnencryptedCredentialDetails struct {
 // StarburstCredentialData represents the data returned by the API for an Starburst credential
 type StarburstCredentialData struct {
 	ID                           *int                                  `json:"id"`
-	AccountID                    int                                   `json:"account_id"`
+	AccountID                    int64                                  `json:"account_id"`
 	ProjectID                    int                                   `json:"project_id"`
 	Type                         string                                `json:"type"`
 	State                        int                                   `json:"state"`
@@ -39,7 +39,7 @@ type StarburstCredentialData struct {
 // It doesn't include the UnencryptedCredentialDetails field which is only returned by the API
 type StarburstCredentialRequest struct {
 	ID                *int                     `json:"id,omitempty"`
-	AccountID         int                      `json:"account_id"`
+	AccountID         int64                    `json:"account_id"`
 	ProjectID         int                      `json:"project_id"`
 	Type              string                   `json:"type"`
 	State             int                      `json:"state"`
