+++ pkg/framework/objects/service_token/model.go
@@ -85,7 +85,7 @@ func ConvertServiceTokenPermissionModelToGrant(
 func ConvertServiceTokenPermissionModelToDataForCreation(
 	ctx context.Context,
 	requiredAllPermissions []ServiceTokenPermission,
-	accountID int,
+	accountID int64,
 ) ([]dbt_cloud.ServiceTokenPermission, diag.Diagnostics) {
 	allPermissionsRequest := make([]dbt_cloud.ServiceTokenPermission, len(requiredAllPermissions))
 	allDiags := diag.Diagnostics{}
@@ -128,7 +128,7 @@ func ConvertServiceTokenPermissionModelToData(
 	ctx context.Context,
 	requiredAllPermissions []ServiceTokenPermission,
 	serviceTokenID int,
-	accountID int,
+	accountID int64,
 ) ([]dbt_cloud.ServiceTokenPermission, diag.Diagnostics) {
 	allPermissionsRequest := make([]dbt_cloud.ServiceTokenPermission, len(requiredAllPermissions))
 	allDiags := diag.Diagnostics{}
